# groupGifts-api-rest/src/main/resources/api.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /admin/{id}:
    get:
      tags:
        - admin-controller
      operationId: getAdmin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDto'
    put:
      tags:
        - admin-controller
      operationId: updateAdmin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDto'
    delete:
      tags:
        - admin-controller
      operationId: removeAdmin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDto'
    patch:
      tags:
        - admin-controller
      operationId: PartialUpdateAdmin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartialUpdateAdminDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDto'
  /admin:
    post:
      tags:
        - admin-controller
      operationId: createAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDto'
  /admin/filter:
    get:
      tags:
        - admin-controller
      operationId: findAdmins
      parameters:
        - name: nick_name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminDto'
  
# USER 
...
components:
  schemas:
...
    AdminDto:
      required:
        - nick_name
        - pass
      type: object
      properties:
        id:
          type: integer
          format: int64
        nick_name:
          type: string
        pass:
          type: string
    UpdateAdminDto:
      required:
        - nick_name
        - pass
      type: object
      properties:
        nick_name:
          type: string
        pass:
          type: string
    CreateAdminDto:
      required:
        - nick_name
        - pass
      type: object
      properties:
        nick_name:
          type: string
        pass:
          type: string
    PartialUpdateAdminDto:
      type: object
      properties:
        nick_name:
          type: string
        pass:
          type: string

# USER
...